+++ PHPUnit/Autoload.php
@@ -90,6 +90,7 @@ spl_autoload_register(
             'phpunit_framework_basetestlistener' => '/Framework/BaseTestListener.php',
             'phpunit_framework_comparator' => '/Framework/Comparator.php',
             'phpunit_framework_comparator_array' => '/Framework/Comparator/Array.php',
+            'phpunit_framework_comparator_datetime' => '/Framework/Comparator/DateTime.php',
             'phpunit_framework_comparator_domdocument' => '/Framework/Comparator/DOMDocument.php',
             'phpunit_framework_comparator_double' => '/Framework/Comparator/Double.php',
             'phpunit_framework_comparator_exception' => '/Framework/Comparator/Exception.php',