+++ PHPUnit/Framework/ComparatorFactory.php
@@ -82,6 +82,7 @@ class PHPUnit_Framework_ComparatorFactory
         $this->register(new PHPUnit_Framework_Comparator_SplObjectStorage);
         $this->register(new PHPUnit_Framework_Comparator_DOMDocument);
         $this->register(new PHPUnit_Framework_Comparator_MockObject);
+        $this->register(new PHPUnit_Framework_Comparator_DateTime);
     }
 
     /**