+++ Tests/Framework/ConstraintTest.php
@@ -923,6 +923,19 @@ Failed asserting that two DOM documents are equal.
 
 EOF
             ),
+            array(
+              new DateTime('2013-03-29 04:13:35', new DateTimeZone('America/New_York')),
+              new DateTime('2013-03-29 04:13:35', new DateTimeZone('America/Chicago')),
+              <<<EOF
+Failed asserting that two DateTimes are equal.
+--- Expected
++++ Actual
+@@ @@
+-2013-03-29T04:13:35-0400
++2013-03-29T04:13:35-0500
+
+EOF
+            ),
         );
     }