+++ ChangeLog.md
@@ -6,6 +6,7 @@ This is the list of changes for the PHPUnit 3.8 release series.
 PHPUnit 3.8.0
 -------------
 
+* Implemented #871: Add Comparator for DateTime objects.
 * Implemented #859: Added PHP label validation to attribute assertions.
 * Implemented #838: Added a base test listener.
 * Implemented #835: Printers that extend `PHPUnit_TextUI_ResultPrinter` should have similar construction